Final data validation; Additional mapping code refactoring
