+++ src/main/groovy/org/opentestsystem/ap/migration/datastore/Migration2673.groovy
@@ -81,39 +81,51 @@ class Migration2673 extends AbstractDataStoreMigration {
     }
 
     private void applyEbsrMigration(ItemEntity itemEntity, Project project) {
+        // Pull import.zip from the items master branch. Expand import.zip to a temporary directory
         Path expandedZip = migrationFileUtil.getExpandedImportZipFile(itemEntity.getItemId(),
                 gitlabClient.getFile(project.getId(), MigrationFileUtil.IMPORT_ZIP_FILENAME, MASTER_BRANCH))
 
         if (expandedZip.toFile().exists()) {
-            //TODO: Identify item directory
-            String itemSourceDir = expandedZip.resolve(String.format("%s-%s-%s", "Item", "200", itemEntity.getItemId()))
-            String itemSyncDir = Files.createTempDirectory(itemEntity.getItemId())
-            String itemFileName = String.format("%s-%s-%s/%s-%s-%s.xml", "Item", "200", itemEntity.getItemId(), "item", "200", itemEntity.getItemId());
-            File itemFile = migrationFileUtil.getFile(itemFileName, expandedZip)
+            // import.zip can contain multiple directories. Identify the imported item source directory
+            String itemSourceDir = migrationFileUtil.getExpandedItemSourceDir(expandedZip, itemEntity.getItemId())
+            // Find the main item file in the source location
+            File itemFile = migrationFileUtil.
+                    findItemFileInPath(expandedZip.resolve(itemSourceDir).toString(), String.format("%s.xml", itemSourceDir))
 
-            final ItemRelease itemRelease = this.appAssembler.getSaaifAssembler()
-                    .readXmlFromFile(itemFile.toPath())
+            if (itemFile != null) {
+                // Create temporary directory to store attachment files that might need synching to S3
+                String itemSyncDir = Files.createTempDirectory(itemEntity.getItemId())
 
-            final IatModelMapper mapper = iatModelMapperFactory.newModelMapper(itemEntity.getItemJson().getType())
+                // Load imported item xml data
+                final ItemRelease itemRelease = this.appAssembler.getSaaifAssembler()
+                        .readXmlFromFile(itemFile.toPath())
 
-            Document itemBody = Jsoup.parseBodyFragment(itemRelease.getItem().getContent().get(0).qti.value, "UTF-8")
-            String qtiFirstParagraph = itemBody.getElementsByTag("p").first().toString()
+                // Create item-specific mapper
+                final IatModelMapper mapper = iatModelMapperFactory.newModelMapper(itemEntity.getItemJson().getType())
 
-            EbsrItem ebsrItem = (EbsrItem) itemEntity.getItemJson()
+                // Load Jsoup document with item Qti content
+                Document itemBody = Jsoup.parseBodyFragment(itemRelease.getItem().getContent().get(0).qti.value, "UTF-8")
 
-            ItemMappingProperties mappingProperties = mapper.mapSingleRichTextContent(qtiFirstParagraph,
-                    ebsrItem,
-                    itemRelease.getItem().getContent().get(0).getApipAccessibility(),
-                    itemSourceDir,
-                    itemSyncDir)
+                if (itemBody.getElementsByTag("p").size() > 0) {
+                    String qtiFirstParagraph = itemBody.getElementsByTag("p").first().toString()
 
+                    EbsrItem ebsrItem = (EbsrItem) itemEntity.getItemJson()
 
-            ebsrItem.getCore().getEn().setPrompt(mappingProperties.getContent())
-            //TODO: add audio and image resources if found in mappingProperties
+                    ItemMappingProperties mappingProperties = mapper.mapSingleRichTextContent(qtiFirstParagraph,
+                            ebsrItem,
+                            itemRelease.getItem().getContent().get(0).getApipAccessibility(),
+                            expandedZip.resolve(itemSourceDir).toString(),
+                            itemSyncDir)
 
-            itemEntity.setItemJson(ebsrItem)
+                    ebsrItem.getCore().getEn().setPrompt(mappingProperties.getContent())
 
-            itemBankSyncManager.syncAttachmentsToDataStore(itemEntity.getItemId(), BRANCH_MASTER, Paths.get(itemSyncDir))
+                    itemEntity.setItemJson(ebsrItem)
+
+                    itemBankSyncManager.syncAttachmentsToDataStore(itemEntity.getItemId(), BRANCH_MASTER, Paths.get(itemSyncDir))
+                }
+            } else {
+                throw new SkipMigration(String.format("Unable to find item file %s for item %s", itemFile.getAbsolutePath(), itemEntity.getId()))
+            }
         }
     }
 }
