+++ src/main/java/org/opentestsystem/ap/migration/mapper/EbsrModelMapper.java
@@ -8,7 +8,7 @@ import org.opentestsystem.ap.common.model.ItemOption;
 import org.opentestsystem.ap.common.saaif.item.ItemRelease;
 import org.opentestsystem.ap.migration.model.ItemMappingProperties;
 import org.opentestsystem.ap.migration.model.ItemMappingResult;
-import org.opentestsystem.ap.migration.util.ImportMapperUtil;
+import org.opentestsystem.ap.migration.util.MigrationMapperUtil;
 
 import java.nio.file.Path;
 import java.util.Arrays;
@@ -28,9 +28,9 @@ public class EbsrModelMapper extends IatModelMapper {
         String itemDestinationFullPath = itemLocalPath.toString();
 
         List<String> answerI = Arrays.asList(
-                ImportMapperUtil.getAttributeValue(release, "itm_att_Answer Key").split(","));
+                MigrationMapperUtil.getAttributeValue(release, "itm_att_Answer Key").split(","));
         List<String> answerII = Arrays.asList(
-                ImportMapperUtil.getAttributeValue(release, "itm_att_Answer Key (Part II)").split(","));
+                MigrationMapperUtil.getAttributeValue(release, "itm_att_Answer Key (Part II)").split(","));
 
         ItemMappingProperties mappingProperties = new ItemMappingProperties();
         for (ItemRelease.Item.Content content : release.getItem().getContent()) {
@@ -42,7 +42,7 @@ public class EbsrModelMapper extends IatModelMapper {
                 ebsrItem.getCore().getEn().setPartAPrompt(mappingProperties.getPartAPrompt());
                 AtomicInteger indexI = new AtomicInteger(0);
                 mappingProperties.getPartAOptions().forEach(option -> {
-                    String currentVal = ImportMapperUtil.getCapitalLetterFromInt(indexI.get());
+                    String currentVal = MigrationMapperUtil.getCapitalLetterFromInt(indexI.get());
                     ebsrItem.getCore().getEn().getPartAOptions()
                             .add(ItemOption.newItemOption(option, "", answerI.contains(currentVal), ""));
                     indexI.getAndIncrement();
@@ -51,7 +51,7 @@ public class EbsrModelMapper extends IatModelMapper {
                 ebsrItem.getCore().getEn().setPartBPrompt(mappingProperties.getPartBPrompt());
                 AtomicInteger indexII = new AtomicInteger(0);
                 mappingProperties.getPartBOptions().forEach(option -> {
-                    String currentVal = ImportMapperUtil.getCapitalLetterFromInt(indexII.get());
+                    String currentVal = MigrationMapperUtil.getCapitalLetterFromInt(indexII.get());
                     ebsrItem.getCore().getEn().getPartBOptions()
                             .add(ItemOption.newItemOption(option, "", answerII.contains(currentVal), ""));
                     indexII.getAndIncrement();
@@ -66,7 +66,7 @@ public class EbsrModelMapper extends IatModelMapper {
                 ebsrItem.getTranslations().getEsp().setPartAPrompt(mappingProperties.getPartAPrompt());
                 AtomicInteger indexI = new AtomicInteger(0);
                 mappingProperties.getPartAOptions().forEach(option -> {
-                    String currentVal = ImportMapperUtil.getCapitalLetterFromInt(indexI.get());
+                    String currentVal = MigrationMapperUtil.getCapitalLetterFromInt(indexI.get());
                     ebsrItem.getTranslations().getEsp().getPartAOptions()
                             .add(ItemOption.newItemOption(option, "", answerI.contains(currentVal), ""));
                     indexI.getAndIncrement();
@@ -75,7 +75,7 @@ public class EbsrModelMapper extends IatModelMapper {
                 ebsrItem.getTranslations().getEsp().setPartBPrompt(mappingProperties.getPartBPrompt());
                 AtomicInteger indexII = new AtomicInteger(0);
                 mappingProperties.getPartBOptions().forEach(option -> {
-                    String currentVal = ImportMapperUtil.getCapitalLetterFromInt(indexII.get());
+                    String currentVal = MigrationMapperUtil.getCapitalLetterFromInt(indexII.get());
                     ebsrItem.getTranslations().getEsp().getPartBOptions()
                             .add(ItemOption.newItemOption(option, "", answerII.contains(currentVal), ""));
                     indexII.getAndIncrement();
