+++ src/main/java/org/opentestsystem/ap/migration/mapper/GiModelMapper.java
@@ -12,9 +12,9 @@ import org.opentestsystem.ap.common.repository.RepositoryUtil;
 import org.opentestsystem.ap.common.saaif.item.ItemRelease;
 import org.opentestsystem.ap.migration.model.ItemMappingProperties;
 import org.opentestsystem.ap.migration.model.ItemMappingResult;
-import org.opentestsystem.ap.migration.util.ImportFileUtil;
-import org.opentestsystem.ap.migration.util.ImportMapperUtil;
-import org.opentestsystem.ap.migration.util.ImportXmlUtil;
+import org.opentestsystem.ap.migration.util.MigrationFileUtil;
+import org.opentestsystem.ap.migration.util.MigrationMapperUtil;
+import org.opentestsystem.ap.migration.util.MigrationXmlUtil;
 
 import java.io.File;
 import java.nio.file.Path;
@@ -67,12 +67,12 @@ public class GiModelMapper extends IatModelMapper {
                                 String sourceDir,
                                 String destinationDir,
                                 ItemMappingResult itemMappingResult) {
-        File gaxFile = ImportMapperUtil.getRendererSpecFileFromItemRelease(release, sourceDir);
+        File gaxFile = MigrationMapperUtil.getRendererSpecFileFromItemRelease(release, sourceDir);
         if (null != gaxFile) {
             copyGaxFiles(gaxFile, sourceDir, destinationDir, itemMappingResult);
 
             if (createGiZipFile(giItem, release, gaxFile, sourceDir, destinationDir, itemMappingResult)) {
-                giItem.getCore().getEn().getContentZipFile().setFileName(ImportFileUtil.GI_SAMPLE_FILE_NAME);
+                giItem.getCore().getEn().getContentZipFile().setFileName(MigrationFileUtil.GI_SAMPLE_FILE_NAME);
                 giItem.getCore().getEn().getContentZipFile().setUploadedDate(new Date());
             }
         }
@@ -86,24 +86,24 @@ public class GiModelMapper extends IatModelMapper {
                                     ItemMappingResult itemMappingResult) {
         try {
             // Create temporary directory to store files that will be included in gi-sample.zip
-            File tempGiSampleDir = ImportFileUtil.createGiSampleDirectory(sourceDir);
+            File tempGiSampleDir = MigrationFileUtil.createGiSampleDirectory(sourceDir);
 
             File destinationDir = new File(destDir);
 
             // Copy gaxFile
-            ImportFileUtil.copyFile(gaxFile, tempGiSampleDir, itemMappingResult);
+            MigrationFileUtil.copyFile(gaxFile, tempGiSampleDir, itemMappingResult);
 
             // Copy all files referenced in the gax file
-            ImportXmlUtil.getFileSpecListFromFile(gaxFile).forEach(fileName -> {
-                ImportFileUtil.copyFile(sourceDir + "/" + fileName,
+            MigrationXmlUtil.getFileSpecListFromFile(gaxFile).forEach(fileName -> {
+                MigrationFileUtil.copyFile(sourceDir + "/" + fileName,
                         tempGiSampleDir.toString(),
                         itemMappingResult);
             });
 
             // Import Qrx file if present in import directory
-            File qrxFile = ImportMapperUtil.getMachineRubricFileFromItemRelease(release, sourceDir);
+            File qrxFile = MigrationMapperUtil.getMachineRubricFileFromItemRelease(release, sourceDir);
             if (null != qrxFile) {
-                ImportFileUtil.copyFile(sourceDir + "/" + qrxFile.getName(),
+                MigrationFileUtil.copyFile(sourceDir + "/" + qrxFile.getName(),
                         tempGiSampleDir.toString() + "/" +
                                 RepositoryUtil.getQrxFileName(giItem.getId(), ItemConstants.ItemVersion.ITEM_VERSION),
                         itemMappingResult);
@@ -111,12 +111,12 @@ public class GiModelMapper extends IatModelMapper {
             }
 
             // Create zip from tempGiSampleDir contents and place it on destinationDir
-            ImportFileUtil.zipDirectory(tempGiSampleDir.toString(),
+            MigrationFileUtil.zipDirectory(tempGiSampleDir.toString(),
                     destinationDir.toString(),
-                    ImportFileUtil.GI_SAMPLE_FILE_NAME);
+                    MigrationFileUtil.GI_SAMPLE_FILE_NAME);
 
             // Delete tempSampleDir
-            ImportFileUtil.deleteDirectory(tempGiSampleDir);
+            MigrationFileUtil.deleteDirectory(tempGiSampleDir);
 
             return true;
         } catch (Exception ex) {
@@ -131,17 +131,17 @@ public class GiModelMapper extends IatModelMapper {
                               String sourceDir,
                               String destinationDir,
                               ItemMappingResult itemMappingResult) {
-        final Document gaxDoc = Jsoup.parse(ImportFileUtil.fileToString(gaxFile),
+        final Document gaxDoc = Jsoup.parse(MigrationFileUtil.fileToString(gaxFile),
                 "", Parser.xmlParser());
 
         gaxDoc.getElementsByTag("Filespec").forEach(filespec -> {
-            ImportFileUtil.copyFile(sourceDir + "/" + filespec.text(),
+            MigrationFileUtil.copyFile(sourceDir + "/" + filespec.text(),
                     destinationDir + "/" + filespec.text(),
                     itemMappingResult);
         });
 
         gaxDoc.getElementsByTag("Image").forEach(image -> {
-            ImportFileUtil.copyFile(sourceDir + "/" + image.attr("src"),
+            MigrationFileUtil.copyFile(sourceDir + "/" + image.attr("src"),
                     destinationDir + "/" + image.attr("src"),
                     itemMappingResult);
         });
