+++ src/main/java/org/opentestsystem/ap/migration/mapper/IatModelMapper.java
@@ -16,8 +16,8 @@ import org.opentestsystem.ap.common.saaif.wordlist.KeywordType;
 import org.opentestsystem.ap.common.saaif.wordlist.WordlistreleaseType;
 import org.opentestsystem.ap.migration.model.*;
 import org.opentestsystem.ap.migration.model.ItemMappingResult;
-import org.opentestsystem.ap.migration.util.ImportFileUtil;
-import org.opentestsystem.ap.migration.util.ImportMapperUtil;
+import org.opentestsystem.ap.migration.util.MigrationFileUtil;
+import org.opentestsystem.ap.migration.util.MigrationMapperUtil;
 
 import java.io.File;
 import java.nio.file.Path;
@@ -609,12 +609,12 @@ public abstract class IatModelMapper {
                               String sourceDir,
                               String destinationDir,
                               ItemMappingResult itemMappingResult) {
-        File qrxFile = ImportMapperUtil.getMachineRubricFileFromItemRelease(release, sourceDir);
+        File qrxFile = MigrationMapperUtil.getMachineRubricFileFromItemRelease(release, sourceDir);
         if (null != qrxFile) {
             String from = sourceDir + "/" + qrxFile.getName();
             String to = destinationDir + "/" +
                     RepositoryUtil.getQrxFileName(item.getId(), ItemConstants.ItemVersion.ITEM_VERSION);
-            ImportFileUtil.copyFile(from, to, itemMappingResult);
+            MigrationFileUtil.copyFile(from, to, itemMappingResult);
             item.getCore().getScoring().setManagedByIat(false);
         }
     }
@@ -632,12 +632,12 @@ public abstract class IatModelMapper {
                              String sourceDir,
                              String destinationDir,
                              ItemMappingResult itemMappingResult) {
-        File eaxFile = ImportMapperUtil.getRendererSpecFileFromItemRelease(release, sourceDir);
+        File eaxFile = MigrationMapperUtil.getRendererSpecFileFromItemRelease(release, sourceDir);
         if (null != eaxFile) {
             String from = sourceDir + "/" + eaxFile.getName();
             String to = destinationDir + "/" +
                     RepositoryUtil.getEaxFileName(item.getId(), ItemConstants.ItemVersion.ITEM_VERSION);
-            ImportFileUtil.copyFile(from, to, itemMappingResult);
+            MigrationFileUtil.copyFile(from, to, itemMappingResult);
             item.getCore().getEquationEditorConfiguration().setManagedByIat(false);
         }
     }
@@ -728,7 +728,7 @@ public abstract class IatModelMapper {
                                          String itemSourceFullPath,
                                          String itemDestinationFullPath,
                                          ItemMappingResult itemMappingResult) {
-        ImportFileUtil.copyFile(itemSourceFullPath + "/" + fileName,
+        MigrationFileUtil.copyFile(itemSourceFullPath + "/" + fileName,
                 itemDestinationFullPath + "/" + fileName,
                 itemMappingResult);
         return createAttachment(fileName);
@@ -752,7 +752,7 @@ public abstract class IatModelMapper {
             item.getAudio().getAudioResources().add(createItemAudioResource(audio, audio.getAudioId()));
 
             audio.getSources().forEach(source ->
-                ImportFileUtil.copyFile(itemSourceFullPath + "/" + source,
+                    MigrationFileUtil.copyFile(itemSourceFullPath + "/" + source,
                         itemDestinationFullPath + "/" + source,
                         importResult));
 
@@ -761,7 +761,7 @@ public abstract class IatModelMapper {
                 item.getCc().setCcRequired("true");
                 item.getCc().getAttachments().add(createAttachment(audio.getTrack()));
                 item.getCc().getItemCcAudioAssociations().add(createItemCCAudioAssociation(audio.getTrack(), audio.getAudioId()));
-                ImportFileUtil.copyFile(itemSourceFullPath + "/" + audio.getTrack(),
+                MigrationFileUtil.copyFile(itemSourceFullPath + "/" + audio.getTrack(),
                         itemDestinationFullPath + "/" + audio.getTrack(),
                         importResult);
             }
@@ -783,7 +783,7 @@ public abstract class IatModelMapper {
         mappingResult.getImageSources().forEach(image -> {
             //int imageId = resourceId.getAndIncrement();
             item.getImages().getImageResources().add(createItemImageResource(image.getSource(), image.getImageId()));
-            ImportFileUtil.copyFile(itemSourceFullPath + "/" + image.getSource(),
+            MigrationFileUtil.copyFile(itemSourceFullPath + "/" + image.getSource(),
                     itemDestinationFullPath + "/" + image.getSource(),
                     importResult);
         });
@@ -1059,7 +1059,7 @@ public abstract class IatModelMapper {
         if (oggFile.exists()) {
             String oggIatFileName = glossaryFileProps.getFileName();
 
-            ImportFileUtil.copyFile(oggFile,
+            MigrationFileUtil.copyFile(oggFile,
                     Paths.get(itemLocalPath.toString(), "glossary", oggIatFileName).toFile(),
                     itemMappingResult);
 
@@ -1068,7 +1068,7 @@ public abstract class IatModelMapper {
             if (m4aFile.exists()) {
                 String m4aIatFileName = m4aFileName;
 
-                ImportFileUtil.copyFile(m4aFile,
+                MigrationFileUtil.copyFile(m4aFile,
                         Paths.get(itemLocalPath.toString(), "glossary", m4aIatFileName).toFile(),
                         itemMappingResult);
             }
@@ -1170,10 +1170,10 @@ public abstract class IatModelMapper {
                                                     String sourceItemFullPath,
                                                     String identifier) {
         List<String> values = new ArrayList<>();
-        File qrxFile = ImportFileUtil.findItemFileInPath(
+        File qrxFile = MigrationFileUtil.findItemFileInPath(
                 sourceItemFullPath, String.format("item_%s_*.qrx", itemId));
         if (qrxFile != null && qrxFile.exists()) {
-            String qrxContent = ImportFileUtil.fileToString(qrxFile);
+            String qrxContent = MigrationFileUtil.fileToString(qrxFile);
             final Document doc = Jsoup.parse(qrxContent, "", Parser.xmlParser());
 
             doc.getElementsByTag("responseDeclaration").forEach(respDeclaration -> {
