+++ src/main/java/org/opentestsystem/ap/migration/mapper/McModelMapper.java
@@ -8,7 +8,7 @@ import org.opentestsystem.ap.common.model.McItem;
 import org.opentestsystem.ap.common.saaif.item.ItemRelease;
 import org.opentestsystem.ap.migration.model.ItemMappingProperties;
 import org.opentestsystem.ap.migration.model.ItemMappingResult;
-import org.opentestsystem.ap.migration.util.ImportMapperUtil;
+import org.opentestsystem.ap.migration.util.MigrationMapperUtil;
 
 import java.nio.file.Path;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -24,7 +24,7 @@ public class McModelMapper extends IatModelMapper {
         McItem mcItem = (McItem) item;
         String itemDestinationFullPath = itemLocalPath.toString();
 
-        String answerValue = ImportMapperUtil.getAttributeValue(release, "itm_att_Answer Key");
+        String answerValue = MigrationMapperUtil.getAttributeValue(release, "itm_att_Answer Key");
 
         ItemMappingProperties mappingProperties = new ItemMappingProperties();
         for (ItemRelease.Item.Content content : release.getItem().getContent()) {
@@ -40,7 +40,7 @@ public class McModelMapper extends IatModelMapper {
                             content.getApipAccessibility(), mappingProperties).getContent();
                     mcItem.getCore().getEn().getOptions().add(
                             ItemOption.newItemOption(optionValue, "",
-                                    answerValue.equals(ImportMapperUtil.getCapitalLetterFromInt(index.get())), rationaleValue));
+                                    answerValue.equals(MigrationMapperUtil.getCapitalLetterFromInt(index.get())), rationaleValue));
                     index.getAndIncrement();
                 }
             } else if (content.getLanguage().equals(ItemConstants.ItemLanguage.LANG_ESN)) {
@@ -57,7 +57,7 @@ public class McModelMapper extends IatModelMapper {
                             content.getApipAccessibility(), mappingProperties).getContent();
                     mcItem.getTranslations().getEsp().getOptions().add(
                             ItemOption.newItemOption(optionValue, "",
-                                    answerValue.equals(ImportMapperUtil.getCapitalLetterFromInt(index.get())), rationaleValue));
+                                    answerValue.equals(MigrationMapperUtil.getCapitalLetterFromInt(index.get())), rationaleValue));
                     index.getAndIncrement();
                 }
             }
