+++ src/main/java/org/opentestsystem/ap/migration/mapper/MiModelMapper.java
@@ -10,7 +10,7 @@ import org.opentestsystem.ap.common.model.Table;
 import org.opentestsystem.ap.common.saaif.item.ItemRelease;
 import org.opentestsystem.ap.migration.model.ItemMappingProperties;
 import org.opentestsystem.ap.migration.model.ItemMappingResult;
-import org.opentestsystem.ap.migration.util.ImportMapperUtil;
+import org.opentestsystem.ap.migration.util.MigrationMapperUtil;
 
 import java.nio.file.Path;
 
@@ -67,7 +67,7 @@ public class MiModelMapper extends IatModelMapper {
                             int colIndex;
                             if (StringUtils.isNumeric(row)) {
                                 rowIndex = Integer.parseInt(row) - 1;
-                                colIndex = ImportMapperUtil.getIntFromLowerCaseLetter(col.charAt(0));
+                                colIndex = MigrationMapperUtil.getIntFromLowerCaseLetter(col.charAt(0));
                                 miItem.getCore().getEn().getTable().getRows()
                                         .get(rowIndex).getCells().get(colIndex).setValue(true);
                             }
