+++ src/main/java/org/opentestsystem/ap/migration/mapper/MsModelMapper.java
@@ -8,7 +8,7 @@ import org.opentestsystem.ap.common.model.MsItem;
 import org.opentestsystem.ap.common.saaif.item.ItemRelease;
 import org.opentestsystem.ap.migration.model.ItemMappingProperties;
 import org.opentestsystem.ap.migration.model.ItemMappingResult;
-import org.opentestsystem.ap.migration.util.ImportMapperUtil;
+import org.opentestsystem.ap.migration.util.MigrationMapperUtil;
 
 import java.nio.file.Path;
 import java.util.ArrayList;
@@ -54,7 +54,7 @@ public class MsModelMapper extends IatModelMapper {
                             content.getApipAccessibility(), mappingProperties).getContent();
                     msItem.getCore().getEn().getOptions().add(
                             ItemOption.newItemOption(optionValue, "",
-                                    answers.contains(ImportMapperUtil.getCapitalLetterFromInt(index.get())), rationaleValue));
+                                    answers.contains(MigrationMapperUtil.getCapitalLetterFromInt(index.get())), rationaleValue));
                     index.getAndIncrement();
                 }
             } else if (content.getLanguage().equals(ItemConstants.ItemLanguage.LANG_ESN)) {
@@ -71,7 +71,7 @@ public class MsModelMapper extends IatModelMapper {
                             content.getApipAccessibility(), mappingProperties).getContent();
                     msItem.getTranslations().getEsp().getOptions().add(
                             ItemOption.newItemOption(optionValue, "",
-                                    answers.contains(ImportMapperUtil.getCapitalLetterFromInt(index.get())), rationaleValue));
+                                    answers.contains(MigrationMapperUtil.getCapitalLetterFromInt(index.get())), rationaleValue));
                     index.getAndIncrement();
                 }
             }
