+++ src/main/java/org/opentestsystem/ap/migration/mapper/TutModelMapper.java
@@ -9,7 +9,7 @@ import org.opentestsystem.ap.common.saaif.metadata.SmarterAppMetadata;
 import org.opentestsystem.ap.migration.model.ItemMappingProperties;
 import org.opentestsystem.ap.migration.model.ItemProps;
 import org.opentestsystem.ap.migration.model.ItemMappingResult;
-import org.opentestsystem.ap.migration.util.ImportFileUtil;
+import org.opentestsystem.ap.migration.util.MigrationFileUtil;
 
 import java.nio.file.Path;
 
@@ -47,7 +47,7 @@ public class TutModelMapper extends IatModelMapper {
 
         //Process Sources
         mappingProperties.getFileSources().forEach(source -> {
-            ImportFileUtil.copyFile(itemSourceFullPath + "/" + source,
+            MigrationFileUtil.copyFile(itemSourceFullPath + "/" + source,
                     itemDestinationFullPath + "/" + source,
                     mappingResult);
         });
