+++ src/main/java/org/opentestsystem/ap/migration/util/MigrationMapperUtil.java
@@ -8,7 +8,7 @@ import java.io.File;
 
 @Slf4j
 @Component
-public class ImportMapperUtil {
+public class MigrationMapperUtil {
 
     private static final char CAPITAL_A = 'A';
 
@@ -38,7 +38,7 @@ public class ImportMapperUtil {
         if (release.getItem().getRendererSpec().size() == 1 &&
                 !release.getItem().getRendererSpec().get(0).getFilename().isEmpty()) {
             String gaxFile = release.getItem().getRendererSpec().get(0).getFilename();
-            return ImportFileUtil.getFileFromDirectory(gaxFile, sourceDir);
+            return MigrationFileUtil.getFileFromDirectory(gaxFile, sourceDir);
         }
         return null;
     }
@@ -47,7 +47,7 @@ public class ImportMapperUtil {
         if (release.getItem().getMachineRubric().size() == 1 &&
                 !release.getItem().getMachineRubric().get(0).getFilename().isEmpty()) {
             String qrxFile = release.getItem().getMachineRubric().get(0).getFilename();
-            return ImportFileUtil.getFileFromDirectory(qrxFile, sourceDir);
+            return MigrationFileUtil.getFileFromDirectory(qrxFile, sourceDir);
         }
         return  null;
     }
