+++ src/main/java/org/opentestsystem/ap/migration/util/MigrationXmlUtil.java
@@ -14,7 +14,7 @@ import java.util.List;
 
 @Slf4j
 @Component
-public class ImportXmlUtil {
+public class MigrationXmlUtil {
 
     public static List<String> getFileSpecListFromFile(File file) {
         List<String> fileList = new ArrayList<>();
